refactor(player): 抽出注册相关逻辑
